requirements.txt
-i https://pypi.org/simple
alembic==1.0.0
aniso8601==3.0.2
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
flask-migrate==2.2.1
flask-restplus==0.11.0
flask-sqlalchemy==2.3.2
flask==1.0.2
idna==2.7
itsdangerous==0.24
jinja2==2.10
jsonschema==2.6.0
mako==1.0.7
markupsafe==1.0
pycparser==2.18
pyjwt==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
six==1.11.0
sqlalchemy==1.2.10
urllib3==1.23; python_version != '3.3.*'
werkzeug==0.14.1
